create inventory and payment grpc client in order service

+++ order/internal/repository/model/order.go
@@ -14,7 +14,7 @@ type Order struct {
 	OrderUuid       string
 	UserUuid        string
 	PartUuids       []string
-	TotalPrice      float64
+	TotalPrice      float32
 	TransactionUuid *string
 	PaymentMethod   *PaymentMethod
 	Status          string
